feat(PageTemplate): rename Static Page to just Page
